requirements.txt
bokeh==2.1.0
contextily==1.0.0
descartes==1.1.0
geopandas==0.7.0
matplotlib==3.2.2
netCDF4==1.5.3
numpy==1.18.5
pandas==1.0.5
pyproj==2.6.1.post1
Shapely==1.7.0
plotly==4.8.1
tqdm==4.46.1
python-magic==0.4.18